Add ContactPointRelationship
This commit adds a new, sub-classed ContactPointRelationship and associated
properties. The point of a contact point relationship is to connect an Agent
contact to an Artifact. The creation of this relationship was born out of
a security use case for cases where a security researcher may need to report
a vulnerability to an Agent but the relationship in the model covers other
use cases such as compliance and general support (see
Core/Vocabularies/ContactPointRelationshipType.md). Additional types of
contact points may also be added in the future.

Altough the issue that sparked this change was tagged is security, this
relationship type was added to Core as it is applicable beyond just
security.

Resolves #861

Signed-off-by: Rose Judge <rose.harber@gmail.com>